page2svg-img: preserve <span> tag
Instead of replacing <svg> element with <img> element, page2svg-img
replaces the <span> parent of the <svg> element. This results in the
suppresion of the <span> element.

This <span> element is important in certain situations (zooming for
example) ; we have to keep it and replace only the <svg> element.

Fixes mathjax/MathJax-node#18